remove imageDownlaod from RequestPerformable & AsyncRequestPerformable + unit tests
