+++ Sources/EZNetworking/Util/Performers/RequestPerformer.swift
@@ -4,7 +4,6 @@ import UIKit
 public protocol RequestPerformable {
     func performTask<T: Decodable>(request: URLRequest, decodeTo decodableObject: T.Type, completion: @escaping((Result<T, NetworkingError>)) -> Void) -> URLSessionDataTask
     func performTask(request: URLRequest, completion: @escaping((VoidResult<NetworkingError>) -> Void)) -> URLSessionDataTask
-    func downloadImageTask(url: URL, completion: @escaping((Result<UIImage, NetworkingError>) -> Void)) -> URLSessionDataTask
 }
 
 public struct RequestPerformer: RequestPerformable {
@@ -50,22 +49,6 @@ public struct RequestPerformer: RequestPerformable {
             } catch {
                 completion(.failure(NetworkingError.unknown))
                 return
-            }
-        }
-    }
-    
-    public func downloadImageTask(url: URL, completion: @escaping((Result<UIImage, NetworkingError>) -> Void)) -> URLSessionDataTask {
-        return urlSession.dataTask(with: url) { data, response, error in
-            do {
-                let validData = try self.urlResponseValidator.validate(data: data, urlResponse: response, error: error)
-                guard let image = UIImage(data: validData) else {
-                    throw NetworkingError.invalidImageData
-                }                
-                completion(.success(image))
-            } catch let networkError as NetworkingError {
-                completion(.failure(networkError))
-            } catch {
-                completion(.failure(.unknown))
             }
         }
     }
